Add files via upload
Completed- Infix to postfix and to prefix both.
What was the error? I wasn't pushing the catenated string, but only one character of the string, so some of the characters were getting left during the conversion of postfix->prefix.
What's left? Evaluation.
More edits for this very version? Yes, in postfix->prefix, memory allocation should be done only if while iterating through the postfix string, an operator is found and also, once an operator is found popping from the stack (that contains only alphanum chars in this part of the code) will return only alphanum chars, so no need to check for the if(alnum()) condition.